fix Oath of Ancients merge
