+++ SolastaUnfinishedBusiness/Translations/en/OathOfAncients-en.txt
@@ -1,0 +1,12 @@
+﻿Condition/&ConditionNaturesWrathDescription=You are entangled in vines preventing movement, disadvantage on dexterity saving throws, and attackers have an advantage. At the end of your turn, make a strength saving throw to escape.
+Condition/&ConditionNaturesWrathTitle=Entangled
+Condition/&ConditionAuraWardingDescription=Gain resistance to natural elements taking half damage from acid, cold, fire, lightning, thunder, and poison.
+Condition/&ConditionAuraWardingTitle=Warded
+Feature/&PowerNaturesWrathDescription=Conjure vines on a target, the target is restrained on a failed strength saving throw. The target rolls a strength saving throwing to escape at the end of their turn.
+Feature/&PowerNaturesWrathTitle=Channel Divinity: Natures Wrath
+Feature/&PowerTurnFaithlessDescription=Perform a prayer, Fey, Fiends, and Elementals that are near must make a wisdom saving throw, on failure they are turned.
+Feature/&PowerTurnFaithlessTitle=Channel Divinity: Turn the Faithless
+Feature/&PowerAuraWardingDescription=Grant you and those around you an ancient ward reducing damage from spells.
+Feature/&PowerAuraWardingTitle=Aura of Warding
+Subclass/&OathOfAncientsDescription=An Oath from a more primitive time which one dedicates themself to utilizing nature for good.
+Subclass/&OathOfAncientsTitle=Oath of the Ancients
