+++ SolastaUnfinishedBusiness/Translations/en/Spells/Spells01-en.txt
@@ -1,27 +1,37 @@
 ﻿Condition/&ConditionMuleDescription=Your speed is not reduced by encumbrance or heavy armor. You can carry 20 additional pounds.
 Condition/&ConditionMuleTitle=Mule
+Condition/&AdditionalConditionEnsnaringStrikDescription=You are entangled in vines preventing movement, disadvantage on dexterity saving throws, and receving piercing damage everyturn. At end of your turn, make a strength throwing save to escape the vines.
+Condition/&AdditionalConditionEnsnaringStrikeTitle=Ensnared
 Feature/&AdditionalDamageSearingSmiteDescription=You ignites in flames.
 Feature/&AdditionalDamageSearingSmiteTitle=Searing Smite
 Feature/&AdditionalDamageThunderousSmiteDescription=Your body cannot stand the the thunder.
 Feature/&AdditionalDamageThunderousSmiteTitle=Thunderous Smite
 Feature/&AdditionalDamageWrathfulSmiteDescription=Your mind explode in pain.
 Feature/&AdditionalDamageWrathfulSmiteTitle=Wrathful Smite
+Feature/&AdditionalDamageEnsnaringStrikeDescription=You are ensnared by vines.
+Feature/&AdditionalDamageEnsnaringStrikeTitle=Ensnaring Strike
 Feedback/&AdditionalDamageSearingSmiteFormat=Searing Smite!
 Feedback/&AdditionalDamageSearingSmiteLine={0} deals more damage to {1} through a searing smite (+{2})
 Feedback/&AdditionalDamageThunderousSmiteFormat=Thunderous Smite!
 Feedback/&AdditionalDamageThunderousSmiteLine={0} deals more damage to {1} through a thunderous smite (+{2})
 Feedback/&AdditionalDamageWrathfulSmiteFormat=Wrathful Smite!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} deals more damage to {1} through a wrathful smite (+{2})
+Feedback/&AdditionalDamageEnsnaringStrikeFormat= Ensnared!
+Feedback/&AdditionalDamageEnsnaringStrikeLine={0} ensnares {1} through a ensnared Strike (+{2})
 Spell/&ChromaticOrbDescription=You hurl a 4-inch-diameter sphere of energy at a creature that you can see within range. You choose acid, cold, fire, lightning, poison, or thunder for the type of orb you create, and then make a ranged spell attack against the target. If the attack hits, the creature takes 3d8 damage of the type you chose.
 Spell/&ChromaticOrbTitle=Chromatic Orb
 Spell/&MuleDescription=The recipient of this spell is able to ignore the effects of heavy loads or armor on movement speed. They can also carry slightly more weight.
 Spell/&MuleTitle=Mule
 Spell/&RadiantMotesDescription=Unleashes a swarm of 4 radiant projectiles that deal 1d4 radiant damage each.\nWhen you cast this spell using a spell slot of 2nd level or higher, the spell creates 1 more projectile for each slot above 1st.
 Spell/&RadiantMotesTitle=Radiant Motes
-Spell/&SearingSmiteDescription=Your next hit causes additional fire damage and your target ignites in flames. At the start of each of its turns until the spell ends, the target must make a Constitution saving throw. On a failed save, it takes 1d6 fire damage. On a successful save, the spells ends.
+Spell/&SearingSmiteDescription=Your next hit causes additional fire damage and your target ignites in flames.
 Spell/&SearingSmiteTitle=Searing Smite
 Spell/&SubSpellChromaticOrbDescription=The creature takes 3d8 {0} damage.
-Spell/&ThunderousSmiteDescription=Your next hit causes additional thunder damage that is audible within 300 feet of you. It must succeed on a Strength saving throw or be pushed 10 feet away from you and knocked prone.
+Spell/&ThunderousSmiteDescription=Your next hit causes additional thunder damage that is audible within 300 feet of you.
 Spell/&ThunderousSmiteTitle=Thunderous Smite
-Spell/&WrathfulSmiteDescription=Your next hit causes additional psychic damage and your target mind explodes in pain. It must make a Wisdom saving throw or be frightened for 1 minute. 
+Spell/&WrathfulSmiteDescription=Your next hit causes additional psychic damage and your target mind explodes in pain.
 Spell/&WrathfulSmiteTitle=Wrathful Smite
+Spell/&EnsnaringStrikeDescription=Your next hit causes additional piercing damage and your target is ensnared by vines.
+Spell/&EnsnaringStrikeTitle=Ensnaring Strike
+Spell/&ConditionEnsnaringStrikeDescription=Your next hit ensnares the foe in vine and causes additional piercing damage if failed a strength saving throw.
+Spell/&ConditionEnsnaringStrikeTitle=Ensnaring Strike
